Only positive maxLength should truncate passwords
This allows us to handle negative maxLength in password inputs.

Specifying a negative maxLength results in the default behavior, which
is to allow an unlimited number of characters[1]. By limiting maxLength
to a minimum of 0, the entire password is used for values of maxLength
less than or equal to 0, while values above 0 use the first maxLength
characters of the password.

[1] https://developer.mozilla.org/en/docs/Web/HTML/Element/input